CastorUtils: Moved mipmap generation algorithm to its own file.
